+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 The changelog lists most feature changes between each release. Search GitHub issues and pull requests for smaller issues.
 
+## [unreleased]
+- change pricing for MOQO provider stadtwerk_tauberfranken
+
 ## 2024-10-01
 - add Cantamen provider teilauto_biberach